+++ sqlite3_web/extra/js-package/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.2
+
+- Add `WebSqlite.close()` to release associated workers.
+
 ## 0.2.1
 
 - Fix internal imports to include `.js` suffix.